fix: restore axon crate layering boundaries
