+++ crates/axon-authz/Cargo.toml
@@ -7,3 +7,6 @@ description = "Axon OAuth scope constants and satisfaction logic"
 license = "MIT"
 
 [dependencies]
+axum = { version = "0.8", features = ["ws"] }
+lab-auth = { git = "https://github.com/jmagar/lab.git", rev = "87cec3241f9baef22335de9cc8629ebbcd8ba047", package = "lab-auth" }
+tracing = "0.1"
