+++ crates/axon-cli/src/commands/sources.rs
@@ -1,9 +1,9 @@
 use axon_core::config::Config;
+use axon_core::env::env_usize_clamped;
 use axon_core::logging::log_info;
 use axon_core::ui::{accent, hyperlink, muted, primary, print_aurora_table};
 use axon_services::system;
 use axon_services::types::Pagination;
-use axon_vector::ops::qdrant::env_usize_clamped;
 use std::error::Error;
 
 pub async fn run_sources(cfg: &Config) -> Result<(), Box<dyn Error>> {
