+++ crates/axon-cli/src/commands/stats.rs
@@ -1,7 +1,6 @@
 use axon_core::config::Config;
 use axon_core::logging::log_info;
 use axon_services::system;
-use axon_vector::ops::stats::display::print_stats_human;
 use std::error::Error;
 
 pub async fn run_stats(cfg: &Config) -> Result<(), Box<dyn Error>> {
@@ -10,7 +9,7 @@ pub async fn run_stats(cfg: &Config) -> Result<(), Box<dyn Error>> {
     if cfg.json_output {
         println!("{}", serde_json::to_string_pretty(&result.payload)?);
     } else {
-        print_stats_human(&result.payload);
+        system::print_stats_human(&result.payload);
     }
     Ok(())
 }
