+++ crates/axon-web/src/server/handlers/rest/auth.rs
@@ -24,11 +24,11 @@ use lab_auth::AuthContext;
 
 pub(crate) fn scope_for_rest_route(method: &str, path: &str) -> Option<&'static str> {
     let scope = match (method, path) {
-        ("GET", p) if p.starts_with("/v1/") => axon_mcp::auth::AxonScope::Read,
-        ("POST", "/v1/query" | "/v1/retrieve" | "/v1/map") => axon_mcp::auth::AxonScope::Read,
-        ("POST", "/v1/dedupe") => axon_mcp::auth::AxonScope::Write,
-        ("POST", p) if p.starts_with("/v1/") => axon_mcp::auth::AxonScope::Write,
-        ("DELETE", p) if p.starts_with("/v1/") => axon_mcp::auth::AxonScope::Write,
+        ("GET", p) if p.starts_with("/v1/") => axon_authz::http::AxonScope::Read,
+        ("POST", "/v1/query" | "/v1/retrieve" | "/v1/map") => axon_authz::http::AxonScope::Read,
+        ("POST", "/v1/dedupe") => axon_authz::http::AxonScope::Write,
+        ("POST", p) if p.starts_with("/v1/") => axon_authz::http::AxonScope::Write,
+        ("DELETE", p) if p.starts_with("/v1/") => axon_authz::http::AxonScope::Write,
         _ => return None,
     };
     Some(scope.as_str())
