+++ crates/axon-web/src/server/handlers/rest_tests.rs
@@ -8,8 +8,8 @@
 
 use super::router;
 use async_trait::async_trait;
+use axon_authz::http::AuthPolicy;
 use axon_jobs::backend::{BackendResult, JobKind, JobPayload};
-use axon_mcp::auth::AuthPolicy;
 use axon_services::context::ServiceContext;
 use axon_services::runtime::ServiceJobRuntime;
 use axon_services::types::ServiceJob;
@@ -172,7 +172,7 @@ fn async_lifecycle_routes_are_declared_for_extract() {
 #[test]
 fn crawl_submit_is_write_scope_across_surfaces() {
     assert_eq!(
-        axon_mcp::auth::scope_for_action("crawl", Some("start")),
+        axon_authz::http::scope_for_action("crawl", Some("start")),
         Some("axon:write")
     );
     assert_eq!(
